local_llm/requirements.txt
Flask==3.0.3
requests
flask-cors==4.0.1
langchain_huggingface
langchain_postgres
langchain-ollama==0.1.3
langchain-core==0.2.38
langchain-text-splitters==0.2.4
psycopg2-binary==2.9.9
numpy==1.26.4
pillow==10.4.0
#for mac user use python3.12.7brewm
#docker exec -it ollama-chatbot ollama pull llama3.2:3b